Moving CBIndices struct to program_factory_common file
